+++ test/lib/registry/main.cpp
@@ -21,7 +21,7 @@ TEST(Lib, Registry) {
     ASSERT_EQ(registry.size<position>(), registry.size<velocity>());
 
     registry.view<position>().each([](auto entity, auto &position) {
-        ASSERT_EQ(position.x, entt::to_integral(entity) + 16);
-        ASSERT_EQ(position.y, entt::to_integral(entity) + 16);
+        ASSERT_EQ(position.x, static_cast<int>(entt::to_integral(entity) + 16));
+        ASSERT_EQ(position.y, static_cast<int>(entt::to_integral(entity) + 16));
     });
 }
